Move the response addons onto a prototype
