+++ lib/middleware/compression.js
@@ -6,6 +6,7 @@ module.exports = function() {
 
 	return function(req, res, next) {
 		let response;
+
 		if(!isHTTP1(req)) {
 			response = Object.create(res, {
 				_implicitHeader: {
