+++ lib/proto/h1.js
@@ -1,0 +1,5 @@
+const { ServerResponse } = require("http");
+const addons = require("./response");
+
+const res = Object.create(ServerResponse.prototype, addons);
+module.exports = res;
