+++ lib/servers.js
@@ -1,7 +1,7 @@
 const proxy = require("./middleware/proxy");
 const errorHandler = require("./middleware/error");
 const fs = require("fs");
-const {markHTTP1, responseAddons} = require("./http1");
+const {markHTTP1} = require("./http1");
 
 exports.create = function(port, options, app) {
 	let servers = [];
@@ -12,6 +12,7 @@ exports.create = function(port, options, app) {
 	if(options.key && options.cert) {
 		let net = require("net");
 		let h2 = require("http2");
+		let responseProto = require("./proto/h2");
 
 		port = Number(port);
 		let httpPort = port + 1;
@@ -23,7 +24,7 @@ exports.create = function(port, options, app) {
 		});
 
 		h2Server.on("request", function(request, response) {
-			Object.assign(response, responseAddons);
+			Object.setPrototypeOf(response, responseProto);
 			app(request, response, handleErrors);
 		});
 
@@ -73,9 +74,10 @@ exports.create = function(port, options, app) {
 			tcpServer.close(onClosed);
 		};
 	} else {
+		let responseProto = require("./proto/h1");
 		let server = require("http").createServer(function(request, response) {
 			markHTTP1(request);
-			Object.assign(response, responseAddons);
+			Object.setPrototypeOf(response, responseProto);
 			app(request, response, handleErrors);
 		});
